+++ src/main/php/inject/Named.class.php
@@ -1,11 +1,13 @@
 <?php namespace inject;
 
+use ArrayAccess, ReturnTypeWillChange;
+
 /**
  * Base class for named lookups
  *
  * @test  xp://inject.unittest.NamedTest
  */
-abstract class Named implements \ArrayAccess {
+abstract class Named implements ArrayAccess {
 
   /**
    * Returns whether this named instance provides a given name
@@ -24,14 +26,18 @@ abstract class Named implements \ArrayAccess {
   public abstract function binding($name);
 
   /** isset() overloading */
+  #[ReturnTypeWillChange]
   public function offsetExists($offset) { return $this->provides($offset); }
 
   /** =[] overloading */
+  #[ReturnTypeWillChange]
   public function offsetGet($offset) { return $this->binding($offset); }
 
   /** []= overloading */
+  #[ReturnTypeWillChange]
   public function offsetSet($offset, $value) { /* Empty */ }
 
   /** unset() overloading */
+  #[ReturnTypeWillChange]
   public function offsetUnset($offset) { /* Empty */ }
 }